% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{add_axes}
\alias{add_axes}
\title{Add axes to a ggplot}
\usage{
add_axes(width = 0.4)
}
\arguments{
\item{width}{Width of the lines. Defaults to 0.4.}
}
\description{
Add axes to a ggplot.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + add_axes()
}
