% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funcs.R
\name{check_args}
\alias{check_args}
\title{check_args}
\usage{
check_args(x, type)
}
\arguments{
\item{x}{a vector}

\item{type}{the type you want to ensure the vector is}
}
\description{
Checks whether `x`, a vector, is of a specific type. Useful for figuring out whether something is a factor / numeric for plotting.
}
\examples{
check_args(c(1,3,4),"numeric")
}
\keyword{cats}
