% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{coord_radar}
\alias{coord_radar}
\title{coord_radar}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{Default is "x".}

\item{start}{Determines where the circle starts. Default is 0.}

\item{direction}{Determines the direction of plotting. Defaults to 1. Change to -1 to go in the opposite direction.}
}
\description{
Coordinate plotting scheme for radar plots.
}
\examples{
mtcars \%>\% rownames_to_column( var = "car" ) \%>\% mutate_at(2:12, funs(scale)) \%>\% sample_n(6) \%>\% gather(mpg:carb, key = "character", value = "value") \%>\% arrange(character) \%>\% ggplot(aes(x=character, y=value, group = car, color = car)) + geom_polygon(fill=NA, aes(color = car), size = 1.5) + geom_polygon(aes(fill = car), fill = NA, size = 2)+ coord_radar()
}
