% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_funcs.R
\name{d_to_r}
\alias{d_to_r}
\title{Convert Cohen's d to r.}
\usage{
d_to_r(d = NULL, n1 = NULL, n2 = NULL)
}
\arguments{
\item{d}{Cohen's d.}

\item{n1}{Sample size of the first group.}

\item{n2}{Sample size of the second group.}
}
\description{
Uses formula found here (https://www.meta-analysis.com/downloads/Meta-analysis%20Converting%20among%20effect%20sizes.pdf) for converting Cohen's d to r.
}
\examples{
d_to_r(0.5, 12, 15)
}
