% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{Explore relationships in your data.}
\usage{
eda(x, n = 99, plot = FALSE)
}
\arguments{
\item{x}{A dataframe, or something that can be coserced to a data frame.}

\item{n}{Number of permutations to use to calcualte a p-value for tau and the ICC. Default is 99 for speed.}

\item{plot}{Return a plot instead of a dataframe? Defaults to FALSE.}
}
\description{
`eda` automates exploring relationships in your dataset. Finds the strength of the relationship between all pairwise combinations of variables so you can quickly find what matters in your data. `eda` results a dataframe (or plot) with each pairwise combination of variables, a test statistic, and p-value calculated by redinglib::permute_tau, redinglib::permute_icc, or `cor.test`. NAs are removed.
}
\examples{
eda(iris)
eda(iris, plot = TRUE)
}
