% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{find_type}
\alias{find_type}
\title{Find the type of a vector.}
\usage{
find_type(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Finds the type of a vector. Most useful in conjunction with plotting functions.
}
\examples{
find_type(1:10)
}
