% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_funcs.R
\name{logodds_to_d}
\alias{logodds_to_d}
\title{Convert log odds ratio to Cohen's D.}
\usage{
logodds_to_d(log_odds = NULL)
}
\arguments{
\item{log_odds}{Logg-odds ratio.}
}
\description{
Uses formula found here (https://www.meta-analysis.com/downloads/Meta-analysis%20Converting%20among%20effect%20sizes.pdf) to convert a log-odds ration to Cohen's d.
}
