% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{move_ticks_inside_x}
\alias{move_ticks_inside_x}
\title{move_ticks_inside_x}
\usage{
move_ticks_inside_x()
}
\description{
Move tick marks inside the x-axis.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + move_ticks_inside_x()
}
