% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{permute_tau}
\alias{permute_tau}
\title{Permute tau.}
\usage{
permute_tau(x, y, n = 99)
}
\arguments{
\item{x}{The first variable.}

\item{y}{The second variable.}

\item{n}{Number of times to permute. Defaults to 99 for speed.}
}
\description{
Permutates tau. Used by redinglib::eda. Returns a p-value.
}
\examples{
find_type(1:10)
}
