% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{plot_cat_relationship}
\alias{plot_cat_relationship}
\title{Plot continous x categorical relationships in your dataset.}
\usage{
plot_cat_relationship(data, categorical_variable)
}
\arguments{
\item{data}{A dataframe.}

\item{categorical_variable}{The bare (unquoted) name of the categorcial variable.}
}
\description{
Plot the relationship between some categorcial variable and each numeric variable.
}
\examples{
plot_cat_relationship(diamonds, clarity)
}
