% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot a bar plot of counts for every non-numeric variable in your dataset.}
\usage{
plot_counts(data)
}
\arguments{
\item{data}{A dataframe.}
}
\description{
Plot a bar plot of counts for every factor/character variable in your dataset.
}
\examples{
plot_counts(diamonds)
}
