% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{plot_histograms}
\alias{plot_histograms}
\title{Plot a histogram for every numeric variable in your dataset.}
\usage{
plot_histograms(data)
}
\arguments{
\item{data}{A dataframe with numeric columns.}
}
\description{
Plot a histogram for every numeric variable in your dataset.
}
\examples{
plot_histograms(mtcars)
}
