% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_funcs.R
\name{r_to_d}
\alias{r_to_d}
\title{Convert correlation coeffeient r to Cohen's d.}
\usage{
r_to_d(r = NULL)
}
\arguments{
\item{r}{Correlation coefficient.}
}
\description{
Uses formula found here (https://www.meta-analysis.com/downloads/Meta-analysis%20Converting%20among%20effect%20sizes.pdf) to convert the correlation coefficient r to Cohen's d.
}
\examples{
r_to_d(0.4)
}
