% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{remove_axes}
\alias{remove_axes}
\title{Remove axes from a ggplot}
\usage{
remove_axes()
}
\description{
Remove axes from a plot.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point() + add_axes()
p + remove_axes()
}
