% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{remove_ticks_x}
\alias{remove_ticks_x}
\title{remove_ticks_x}
\usage{
remove_ticks_x()
}
\description{
Remove tick marks from the x-axis.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + remove_ticks_x()
}
