% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{rotate_labels}
\alias{rotate_labels}
\title{Rotate x-axis labels}
\usage{
rotate_labels(angle = 45)
}
\arguments{
\item{angle}{Angle to rotate the labels. Defaults to 45.}
}
\description{
Rotates x-axis labels on a ggplot.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + rotate_labels()
}
