% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_color_carto}
\alias{scale_color_carto}
\title{Carto color scale}
\usage{
scale_color_carto(...)
}
\arguments{
\item{...}{Additional arguments passed to ggplot2::discrete_scale.}
}
\description{
Carto charts (https://carto.com/carto-colors/) color scale for ggplot graphs.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, color = factor(cyl))) + geom_point()
p + scale_color_carto()
}
