% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_color_OkabeIto}
\alias{scale_color_OkabeIto}
\title{OkabeIto color scale}
\usage{
scale_color_OkabeIto(...)
}
\arguments{
\item{...}{Additional arguments passed to ggplot2::discrete_scale.}
}
\description{
OkabeIto color scale for ggplot graphs.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, color = factor(cyl))) + geom_point()
p + scale_color_OkabeIto()
}
