% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_carto}
\alias{scale_fill_carto}
\title{Carto fill scale}
\usage{
scale_fill_carto(...)
}
\arguments{
\item{...}{Additional arguments passed to ggplot2::discrete_scale.}
}
\description{
Carto charts (https://carto.com/carto-colors/) fill scale for ggplot graphs.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, fill = factor(cyl))) + geom_point(shape = 22)
p + scale_fill_carto()
}
