% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_OkabeIto}
\alias{scale_fill_OkabeIto}
\title{OkabeIto fill scale}
\usage{
scale_fill_OkabeIto(...)
}
\arguments{
\item{...}{Additional arguments passed to ggplot2::discrete_scale.}
}
\description{
OkabeIto fill scale for ggplot graphs.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, fill = factor(cyl))) + geom_point(shape = 22)
p + scale_fill_OkabeIto()
}
