% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_world}
\alias{scale_fill_world}
\title{World fill scale}
\usage{
scale_fill_world(...)
}
\arguments{
\item{...}{Additional arguments passed to ggplot2::discrete_scale.}
}
\description{
'World' fill scale for ggplot graphs.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, fill = factor(cyl))) + geom_point(shape = 22)
p + scale_fill_world()
}
