% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{theme_simple}
\alias{theme_simple}
\title{A simple plotting theme.}
\usage{
theme_simple(font_size = 14, font_family = "", line_size = 0.4)
}
\arguments{
\item{font_size}{Base size of the font. Default is 14.}

\item{font_family}{Font family.}

\item{line_size}{Size of axis ticks. Defauts to 0.4.}
}
\description{
My preferred plotting theme. No gridlines and larger axis labels.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + theme_simple()
}
